+++ src/task/src/TaskExecutor.php
@@ -14,6 +14,7 @@ namespace Hyperf\Task;
 
 use Hyperf\Task\Exception\TaskException;
 use Hyperf\Task\Exception\TaskExecuteException;
+use Hyperf\Utils\Serializer\ExceptionNormalizer;
 use Swoole\Server;
 
 class TaskExecutor
@@ -29,13 +30,19 @@ class TaskExecutor
     protected $factory;
 
     /**
+     * @var ExceptionNormalizer
+     */
+    protected $normalizer;
+
+    /**
      * @var bool
      */
     protected $isTaskEnvironment = false;
 
-    public function __construct(ChannelFactory $factory)
+    public function __construct(ChannelFactory $factory, ExceptionNormalizer $normalizer)
     {
         $this->factory = $factory;
+        $this->normalizer = $normalizer;
     }
 
     public function setServer(Server $server): void
@@ -53,7 +60,14 @@ class TaskExecutor
             throw new TaskExecuteException('Task execute failed.');
         }
 
-        return $this->factory->pop($taskId, $timeout);
+        $result = $this->factory->pop($taskId, $timeout);
+
+        if ($result instanceof Exception) {
+            $exception = $this->normalizer->denormalize($result->attributes, $result->class);
+            throw $exception;
+        }
+
+        return $result;
     }
 
     public function isTaskEnvironment(): bool